refactor(functions): dlsite endpointを薄いハンドラ+オーケストレータ構成へ再配置 (SPR-231 段階②)
931行に肥大していたdlsite-individual-info-api.tsをendpoints/dlsite/の
4ファイルへ分割し、ロジックの所在を予測可能にする（挙動不変・移動中心）:

- fetch-dlsite-unified-data.ts: 薄いハンドラ（CloudEvent unwrap→
  オーケストレータ呼び出し→結果ログ）
- run-unified-data-collection.ts: オーケストレータ（サイクル解決→
  バッチループ→完了処理→creator recompute drain）
- collection-metadata.ts: run metadataと継続cursorのライフサイクル
- process-batch.ts: 1バッチ分の実行と集計/ログヘルパ

リネームはexecuteUnifiedDataCollection→runUnifiedDataCollectionの1件のみ。
ハンドラのエクスポート名（Cloud Functions登録名）は不変。テストは
endpoints/dlsite/__tests__/へ移設し、変更はimportパス深さと
endpoint import分割のみ（アサーション無変更＝挙動不変の合格条件）。
旧ファイル名を参照していたコメント・docsリンク（database-schema.md、
lint:docsが検出）も追従。
