+++ apps/functions/src/endpoints/dlsite/run-unified-data-collection.ts
@@ -1,47 +1,49 @@
 /**
- * DLsite 統合データ収集エンドポイント
+ * DLsite 統合データ収集: オーケストレータ（この関数群が run の本処理）
  *
- * 新しいshared-types構造と薄いマッパーを使用したバージョン
- * レガシーフィールドを段階的に削除
+ * サイクル解決（継続 or 新規・due-set 計算・fast-lane 並べ替え）→ バッチループ →
+ * 完了処理 → creator recompute drain の流れを組み立てる。
+ * ハンドラ（fetch-dlsite-unified-data.ts）から `runUnifiedDataCollection` が呼ばれる。
  */
 
-import type { CloudEvent } from "@google-cloud/functions-framework";
-import type {
-	CollectionMetadata,
-	DLsiteApiResponse,
-	WorkDocument,
-} from "@suzumina.click/shared-types";
-import { Timestamp } from "../infrastructure/database/firestore";
-import { recomputeCreatorStats } from "../services/dlsite/creator-firestore";
+import type { CollectionMetadata, WorkDocument } from "@suzumina.click/shared-types";
+import { Timestamp } from "../../infrastructure/database/firestore";
+import { recomputeCreatorStats } from "../../services/dlsite/creator-firestore";
 import {
 	resetCreatorRecomputeQueue,
 	takeQueuedCreators,
-} from "../services/dlsite/creator-recompute-queue";
-import { getExistingWorksMap } from "../services/dlsite/dlsite-firestore";
+} from "../../services/dlsite/creator-recompute-queue";
+import { getExistingWorksMap } from "../../services/dlsite/dlsite-firestore";
 import {
 	getDlsiteReadMetrics,
 	resetDlsiteReadMetrics,
-} from "../services/dlsite/dlsite-read-metrics";
-import { batchFetchIndividualInfo } from "../services/dlsite/individual-info-api-client";
-import {
-	type ProcessingResult,
-	processBatchUnifiedDLsiteData,
-} from "../services/dlsite/unified-data-processor";
-import { collectWorkIdsForProduction } from "../services/dlsite/work-id-collector";
-import { orderNewWorksFirst } from "../services/dlsite/work-ordering";
+} from "../../services/dlsite/dlsite-read-metrics";
+import { collectWorkIdsForProduction } from "../../services/dlsite/work-id-collector";
+import { orderNewWorksFirst } from "../../services/dlsite/work-ordering";
 import {
 	classifyAndFilterStableTier,
 	classifyWorkTiers,
 	getStableCandidateIds,
 	type TieredWorkIds,
 	toDueWorkIds,
-} from "../services/dlsite/work-tiering";
-import { bulkCheckPriceHistoryExistsToday, getJSTDate } from "../services/price-history";
-import { chunkArray } from "../shared/array-utils";
-import { withClearedUndefined } from "../shared/firestore-write";
-import * as logger from "../shared/logger";
-import { decodePubsubMode, type MessagePublishedData } from "../shared/pubsub-utils";
-import { createRunMetadataStore } from "../shared/run-metadata";
+} from "../../services/dlsite/work-tiering";
+import { bulkCheckPriceHistoryExistsToday, getJSTDate } from "../../services/price-history";
+import { chunkArray } from "../../shared/array-utils";
+import * as logger from "../../shared/logger";
+import {
+	finalizeCompletedProcessing,
+	getContinuationInfo,
+	getOrCreateUnifiedMetadata,
+	initializeNewBatchProcessing,
+	updateUnifiedMetadata,
+} from "./collection-metadata";
+import {
+	BATCH_SIZE,
+	type BatchProcessingInfo,
+	MAX_EXECUTION_TIME,
+	processBatch,
+	type UnifiedFetchResult,
+} from "./process-batch";
 
 /**
  * SPR-229 Stage②: ティア差分によるdue-setフィルタの有効/無効フラグ。
@@ -52,389 +54,6 @@ import { createRunMetadataStore } from "../shared/run-metadata";
  */
 function isTierFilteringEnabled(): boolean {
 	return process.env.DLSITE_TIER_FILTERING_ENABLED !== "false";
-}
-
-// 統合メタデータ保存用の定数
-const UNIFIED_METADATA_DOC_ID = "unified_data_collection_metadata";
-const METADATA_COLLECTION = "dlsiteMetadata";
-
-// バッチ処理設定
-const MAX_CONCURRENT_API_REQUESTS = 5; // 6 → 3 → 5（バッチ処理を効率化: 50件を10回の並列処理で実行）
-const API_REQUEST_DELAY = 400;
-const BATCH_SIZE = 50; // 100 → 50に削減（エラー率低下とタイムアウト回避）
-const MAX_EXECUTION_TIME = 270000; // 4.5分（Cloud Functions 5分タイムアウトより短く設定）
-
-// Note: CollectionMetadata type is now imported from @suzumina.click/shared-types
-
-/**
- * 統合処理結果の型定義
- */
-interface UnifiedFetchResult {
-	workCount: number;
-	apiCallCount: number;
-	basicDataUpdated: number;
-	error?: string;
-	unificationComplete?: boolean;
-}
-
-/**
- * バッチ処理情報の型定義
- */
-interface BatchProcessingInfo {
-	batchNumber: number;
-	totalBatches: number;
-	workIds: string[];
-	startTime: Timestamp;
-	existingWorksMap?: Map<string, WorkDocument>;
-	/** SPR-229: 週次フルスイープ時、通常runならstableとしてスキップされていたはずの作品ID（取りこぼし検知用） */
-	wouldSkipStableIds?: Set<string>;
-}
-
-/**
- * 統合データ収集メタデータのストア（SPR-231: 骨格は shared/run-metadata に集約）
- *
- * update 時は undefined フィールドを FieldValue.delete() へ変換してから書き込む。
- * ignoreUndefinedProperties 環境では undefined がスキップされ旧値が残る（sticky）ため、
- * メタデータをクリア可能にする（`withClearedUndefined` 参照）。
- */
-const unifiedMetadataStore = createRunMetadataStore<CollectionMetadata>({
-	collection: METADATA_COLLECTION,
-	docId: UNIFIED_METADATA_DOC_ID,
-	createInitial: () => ({
-		lastFetchedAt: Timestamp.now(),
-		isInProgress: false,
-		unifiedSystemStarted: Timestamp.now(),
-		migrationVersion: "v2",
-	}),
-	sanitizeUpdate: withClearedUndefined,
-});
-
-/**
- * 統合データ収集メタデータの取得または初期化
- */
-async function getOrCreateUnifiedMetadata(): Promise<CollectionMetadata> {
-	return unifiedMetadataStore.getOrCreate();
-}
-
-/**
- * 統合データ収集メタデータの更新
- */
-async function updateUnifiedMetadata(update: Partial<CollectionMetadata>): Promise<void> {
-	await unifiedMetadataStore.update(update);
-}
-
-/**
- * APIレスポンスがない場合のエラーレスポンスを作成
- */
-function createNoResponseResult(workIdsLength: number): UnifiedFetchResult {
-	return {
-		workCount: 0,
-		apiCallCount: workIdsLength,
-		basicDataUpdated: 0,
-		error: "APIレスポンスなし",
-	};
-}
-
-/**
- * 成功した処理結果をカウント
- */
-function countSuccessfulUpdates(result: ProcessingResult) {
-	if (!result.success || !result.updates) {
-		return { work: 0, circle: 0, creator: 0, priceHistory: 0 };
-	}
-
-	return {
-		work: result.updates.work ? 1 : 0,
-		circle: result.updates.circle ? 1 : 0,
-		creator: result.updates.creators ? 1 : 0,
-		priceHistory: result.updates.priceHistory ? 1 : 0,
-	};
-}
-
-/**
- * 処理結果を集計
- */
-function aggregateProcessingResults(processingResults: ProcessingResult[]): {
-	workUpdated: number;
-	circleUpdated: number;
-	creatorUpdated: number;
-	priceHistoryUpdated: number;
-	errors: string[];
-} {
-	let workUpdated = 0;
-	let circleUpdated = 0;
-	let creatorUpdated = 0;
-	let priceHistoryUpdated = 0;
-	const errors: string[] = [];
-
-	for (const result of processingResults) {
-		const counts = countSuccessfulUpdates(result);
-		workUpdated += counts.work;
-		circleUpdated += counts.circle;
-		creatorUpdated += counts.creator;
-		priceHistoryUpdated += counts.priceHistory;
-
-		if (result.errors?.length > 0) {
-			errors.push(...result.errors);
-		}
-	}
-
-	return {
-		workUpdated,
-		circleUpdated,
-		creatorUpdated,
-		priceHistoryUpdated,
-		errors,
-	};
-}
-
-/**
- * バッチ処理結果をUnifiedFetchResult形式に変換
- */
-function createBatchResult(
-	aggregatedResults: {
-		workUpdated: number;
-		circleUpdated: number;
-		creatorUpdated: number;
-		errors: string[];
-	},
-	workIdsLength: number,
-): UnifiedFetchResult {
-	return {
-		workCount: aggregatedResults.workUpdated,
-		apiCallCount: workIdsLength,
-		basicDataUpdated: aggregatedResults.workUpdated,
-		unificationComplete: aggregatedResults.errors.length === 0,
-	};
-}
-
-/**
- * API呼び出しの失敗をログ出力
- */
-function logApiFailures(batchNumber: number, failedWorkIds: string[]): void {
-	if (failedWorkIds.length > 0) {
-		logger.warn(`バッチ ${batchNumber}: ${failedWorkIds.length}件の取得失敗`);
-		logger.debug(
-			`失敗ID一覧: ${failedWorkIds.slice(0, 10).join(", ")}${failedWorkIds.length > 10 ? "..." : ""}`,
-		);
-	}
-}
-
-/**
- * 価格履歴の保存状況をデバッグログ出力
- */
-function logPriceHistoryDebug(
-	batchNumber: number,
-	priceHistoryUpdated: number,
-	apiResponses: DLsiteApiResponse[],
-	existingWorksMap?: Map<string, WorkDocument>,
-): void {
-	if (priceHistoryUpdated === 0 && apiResponses.length > 0) {
-		logger.warn(`[DEBUG] バッチ ${batchNumber}: 価格履歴が1件も保存されませんでした`);
-		logger.warn(`[DEBUG] 現在時刻: ${new Date().toISOString()}`);
-		logger.warn(
-			`[DEBUG] JST時刻: ${new Date().toLocaleString("ja-JP", { timeZone: "Asia/Tokyo" })}`,
-		);
-		// 既存作品の状態を確認
-		const sampleWorkId = apiResponses[0]?.workno;
-		if (sampleWorkId && existingWorksMap) {
-			const existingWork = existingWorksMap.get(sampleWorkId);
-			logger.warn(
-				`[DEBUG] サンプル作品 ${sampleWorkId} の既存データ: ${existingWork ? "あり" : "なし"}`,
-			);
-		}
-	}
-}
-
-/**
- * バッチ処理の結果をログ出力
- */
-function logBatchComplete(
-	batchNumber: number,
-	workIdsLength: number,
-	apiDataMapSize: number,
-	aggregatedResults: ReturnType<typeof aggregateProcessingResults>,
-): void {
-	logger.info(`バッチ ${batchNumber}: 統合処理完了`, {
-		入力: workIdsLength,
-		API成功: apiDataMapSize,
-		Work更新: aggregatedResults.workUpdated,
-		Circle更新: aggregatedResults.circleUpdated,
-		Creator更新: aggregatedResults.creatorUpdated,
-		価格履歴: aggregatedResults.priceHistoryUpdated,
-		エラー数: aggregatedResults.errors.length,
-	});
-}
-
-/**
- * SPR-229: 週次フルスイープの取りこぼし検知ログ
- *
- * `wouldSkipStableIds`（通常runならstableとしてスキップされていたはずの作品）のうち、
- * 実際にWorkの変化が検出された件数を出す。0件が理想（stableの90日窓が妥当であることの裏付け）。
- * 非0が続く場合はStage③で`VOLATILE_RELEASE_WINDOW_DAYS`の見直しを検討する。
- */
-function logFullSweepMissedChanges(
-	batchNumber: number,
-	processingResults: ProcessingResult[],
-	wouldSkipStableIds: Set<string> | undefined,
-): void {
-	if (!wouldSkipStableIds || wouldSkipStableIds.size === 0) {
-		return;
-	}
-
-	const targeted = processingResults.filter((r) => wouldSkipStableIds.has(r.workId));
-	const missed = targeted.filter((r) => r.updates.work);
-
-	logger.info(`バッチ ${batchNumber}: 週次フルスイープ取りこぼし検知`, {
-		stable想定件数: targeted.length,
-		変化検出件数: missed.length,
-		...(missed.length > 0 && { 変化検出workIds: missed.map((r) => r.workId) }),
-	});
-}
-
-/**
- * バッチ処理実行
- */
-async function processBatch(batchInfo: BatchProcessingInfo): Promise<UnifiedFetchResult> {
-	const { batchNumber, totalBatches, workIds, existingWorksMap, wouldSkipStableIds } = batchInfo;
-
-	logger.info(`バッチ ${batchNumber}/${totalBatches} 処理開始: ${workIds.length}件`);
-
-	try {
-		// 1. Individual Info API 呼び出し（バッチ）
-		const { results: apiDataMap, failedWorkIds } = await batchFetchIndividualInfo(workIds, {
-			maxConcurrent: MAX_CONCURRENT_API_REQUESTS,
-			batchDelay: API_REQUEST_DELAY,
-		});
-
-		if (apiDataMap.size === 0) {
-			logger.warn(`バッチ ${batchNumber}: APIレスポンスなし`);
-			return createNoResponseResult(workIds.length);
-		}
-
-		// 失敗作品のログ
-		logApiFailures(batchNumber, failedWorkIds);
-
-		const apiResponses = Array.from(apiDataMap.values());
-
-		// デバッグ: API取得数とデータ内容を確認
-		logger.info(`[DEBUG] バッチ ${batchNumber}: API取得成功数=${apiResponses.length}`);
-		if (apiResponses.length > 0 && apiResponses[0]) {
-			logger.info(`[DEBUG] サンプルworkno: ${apiResponses[0].workno}`);
-		}
-
-		// 2. 新しい統合処理を使用
-		const processingResults = await processBatchUnifiedDLsiteData(apiResponses, {
-			skipPriceHistory: false, // 価格履歴も含めて全て更新
-			forceUpdate: false, // 差分チェックあり
-			existingWorksMap, // 既存作品マップを渡す
-		});
-
-		// 3. 結果の集計
-		const aggregatedResults = aggregateProcessingResults(processingResults);
-
-		// デバッグ: 価格履歴保存の詳細
-		logPriceHistoryDebug(
-			batchNumber,
-			aggregatedResults.priceHistoryUpdated,
-			apiResponses,
-			existingWorksMap,
-		);
-
-		// ログ出力
-		logBatchComplete(batchNumber, workIds.length, apiDataMap.size, aggregatedResults);
-		logFullSweepMissedChanges(batchNumber, processingResults, wouldSkipStableIds);
-
-		return createBatchResult(aggregatedResults, workIds.length);
-	} catch (error) {
-		logger.error(`バッチ ${batchNumber} 処理エラー:`, { error });
-		return {
-			workCount: 0,
-			apiCallCount: workIds.length,
-			basicDataUpdated: 0,
-			error: error instanceof Error ? error.message : "不明なエラー",
-		};
-	}
-}
-
-/**
- * バッチ処理の継続情報を取得
- */
-function getContinuationInfo(
-	metadata: CollectionMetadata,
-): { isContinuation: true; allWorkIds: string[]; startBatch: number } | { isContinuation: false } {
-	if (
-		metadata.batchProcessingMode &&
-		metadata.allWorkIds &&
-		metadata.currentBatch !== undefined &&
-		metadata.totalBatches !== undefined
-	) {
-		// currentBatchがtotalBatchesに達している場合は新規処理として扱う
-		if ((metadata.currentBatch ?? 0) >= metadata.totalBatches) {
-			logger.info("前回の処理は完了済み。新規処理として開始します", {
-				currentBatch: metadata.currentBatch,
-				totalBatches: metadata.totalBatches,
-			});
-			return { isContinuation: false as const };
-		}
-
-		logger.info("バッチ処理を継続します", {
-			currentBatch: metadata.currentBatch,
-			totalBatches: metadata.totalBatches,
-			allWorkIdsLength: metadata.allWorkIds?.length || 0,
-		});
-
-		return {
-			isContinuation: true as const,
-			allWorkIds: metadata.allWorkIds,
-			startBatch: metadata.currentBatch,
-		};
-	}
-
-	return { isContinuation: false as const };
-}
-
-/**
- * 新規バッチ処理を初期化
- *
- * @param tierBreakdown SPR-229: このサイクルのティア別内訳（観測用。省略時=ティアリング無効）
- * @param isFullSweepCycle SPR-229: 週次フルスイープ実行かどうか
- * @param fullSweepWouldSkipWorkIds SPR-229: 週次フルスイープ時、通常runならスキップされていた作品ID（取りこぼし検知用）
- */
-async function initializeNewBatchProcessing(
-	allWorkIds: string[],
-	batches: string[][],
-	tierBreakdown?: CollectionMetadata["tierBreakdown"],
-	isFullSweepCycle = false,
-	fullSweepWouldSkipWorkIds?: string[],
-): Promise<void> {
-	logger.info(
-		`新規バッチ処理開始: 総作品数=${allWorkIds.length}, バッチ数=${batches.length}, バッチサイズ=${BATCH_SIZE}`,
-	);
-	logger.info(
-		`処理時間制限: ${MAX_EXECUTION_TIME / 1000}秒, 予想処理時間: ${batches.length * 30}秒（30秒/バッチ）`,
-	);
-
-	await updateUnifiedMetadata({
-		isInProgress: true,
-		batchProcessingMode: true,
-		allWorkIds,
-		totalBatches: batches.length,
-		currentBatch: 0,
-		totalWorks: allWorkIds.length,
-		processedWorks: 0,
-		basicDataUpdated: 0,
-		completedBatches: [],
-		lastError: undefined,
-		currentBatchStartTime: Timestamp.now(),
-		migrationVersion: "v2",
-		tierBreakdown,
-		isFullSweepCycle,
-		fullSweepWouldSkipWorkIds: isFullSweepCycle ? fullSweepWouldSkipWorkIds : undefined,
-		// SPR-229: 新規サイクル開始時点で pending 状態は解消済み（effectiveFullSweep に
-		// 反映済み、または今回対象外）のため必ずクリアする。
-		pendingFullSweep: undefined,
-	});
 }
 
 /**
@@ -580,34 +199,6 @@ async function executeBatchLoop(
 	}
 
 	return { totalUpdated, totalApiCalls, completedBatches, allBatchesCompleted };
-}
-
-/**
- * 処理完了時のメタデータ更新
- */
-async function finalizeCompletedProcessing(
-	allWorkIds: string[],
-	totalUpdated: number,
-): Promise<void> {
-	logger.info("全バッチ処理完了", {
-		totalWorks: allWorkIds.length,
-		totalUpdated,
-	});
-
-	await updateUnifiedMetadata({
-		isInProgress: false,
-		lastSuccessfulCompleteFetch: Timestamp.now(),
-		lastFetchedAt: Timestamp.now(),
-		batchProcessingMode: false,
-		currentBatch: undefined,
-		totalBatches: undefined,
-		allWorkIds: undefined,
-		completedBatches: undefined,
-		// SPR-229: サイクル終了時にティア/週次フルスイープ関連の一時フィールドをクリアする
-		tierBreakdown: undefined,
-		isFullSweepCycle: undefined,
-		fullSweepWouldSkipWorkIds: undefined,
-	});
 }
 
 /**
@@ -799,7 +390,9 @@ export async function resolveCycleInfo(
  *
  * @param forceFullSweep SPR-229 Stage②: 週次フルスイープ実行時 true（ティア差分を無視して全件を対象にする）
  */
-async function executeUnifiedDataCollection(forceFullSweep = false): Promise<UnifiedFetchResult> {
+export async function runUnifiedDataCollection(
+	forceFullSweep = false,
+): Promise<UnifiedFetchResult> {
 	// SPR-225 Stage 0/P0: dlsite 同期 reads の内訳をこの run の分だけ計測する（挙動は変えない）。
 	resetDlsiteReadMetrics();
 	// SPR-225 Stage 1: 変更 creator の recompute キューを run 開始でクリアする。
@@ -887,45 +480,5 @@ async function executeUnifiedDataCollection(forceFullSweep = false): Promise<Uni
 		// creator-sync / recompute を網羅）。recompute(drain) の読み取りも計上するため recompute の
 		// 後に出す。Cloud Monitoring の絞り込みキーにするため、メッセージは Stage を含めない恒久キー。
 		logger.info("dlsite reads計測(run)", { ...getDlsiteReadMetrics() });
-	}
-}
-
-/**
- * Cloud Functions エントリーポイント
- */
-export async function fetchDLsiteUnifiedData(
-	event: CloudEvent<MessagePublishedData>,
-): Promise<void> {
-	const mode = decodePubsubMode(event.data);
-	const isWeeklyFullSweep = mode === "weekly_full_sweep";
-
-	logger.info("統合データ収集開始", {
-		eventType: event.type,
-		mode,
-		週次フルスイープ: isWeeklyFullSweep,
-		timestamp: new Date().toISOString(),
-		timezone: Intl.DateTimeFormat().resolvedOptions().timeZone,
-	});
-
-	try {
-		const result = await executeUnifiedDataCollection(isWeeklyFullSweep);
-
-		if (result.error) {
-			logger.error(`統合データ収集エラー: ${result.error}`);
-		} else {
-			logger.info("統合データ収集完了", {
-				更新作品数: result.basicDataUpdated,
-				API呼び出し数: result.apiCallCount,
-				統合完了: result.unificationComplete,
-			});
-		}
-	} catch (error) {
-		logger.error("予期しないエラー:", error);
-
-		await updateUnifiedMetadata({
-			isInProgress: false,
-			lastError: error instanceof Error ? error.message : "不明なエラー",
-			lastFetchedAt: Timestamp.now(),
-		});
 	}
 }
